+++ bindings/nodejs/lib/types/block/payload/candidacy-announcement.ts
