+++ bindings/nodejs/lib/types/block/payload/signed_transaction/index.ts
@@ -3,4 +3,4 @@
 
 export * from './transaction';
 export * from './unlock';
-export * from './signed_transaction';
+export * from './signed-transaction';
