+++ bindings/nodejs/lib/types/block/payload/signed_transaction/signed-transaction.ts
